refactor: 반응형 작업을..위한 리팩토링 진행 (오류 수정중)
