+++ src/app/write/_components/WriteHeader.tsx
@@ -1,0 +1,8 @@
+const WriteHeader = () => (
+    <div className="space-y-2 pb-10">
+      <h1 className="text-title1 font-bold leading-[150%] text-text-04">게시물 작성하기</h1>
+      <p className="text-title2 text-text-03">나만의 룩이나 소개하고 싶은 옷을 공유해보세요!</p>
+    </div>
+  );
+  
+  export default WriteHeader;
